+++ ktorfit-ksp/src/main/kotlin/de/jensklingenberg/ktorfit/model/ClassData.kt
@@ -49,6 +49,7 @@ fun KSClassDeclaration.toClassData(logger: KSPLogger): ClassData {
             "io.ktor.http.URLBuilder",
             "io.ktor.http.takeFrom",
             "io.ktor.http.decodeURLQueryComponent",
+            annotationsAttributeKey.packageName + "." + annotationsAttributeKey.name,
             typeDataClass.packageName + "." + typeDataClass.name,
         )