+++ docs/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project orients towards [Semantic Versioning](http://semver.org/spec/v2
 Note: This project needs KSP to work and every new Ktorfit with an update of the KSP version is technically a breaking change.
 But there is no intent to bump the Ktorfit major version for every KSP update. 
 
+# [Unreleased]()
+* Supported Kotlin version:
+* Supported KSP version:
+* Ktor version:
+
+## Added
+- Include function annotations in request attribute
+See https://foso.github.io/Ktorfit/requests/#annotations
+
 # [2.2.0]()
 * Supported Kotlin version: 2.0.0; 2.0.10; 2.0.20, 2.1.0-Beta1; 2.0.21-RC, 2.0.21, 2.1.0-RC, 2.1.0-RC2, 2.1.0
 * Supported KSP version: 1.0.27, 1.0.28, 1.0.29